backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
PyJWT>=2.0.0
email-validator>=2.0.0
